+++ bencher/__init__.py
@@ -17,4 +17,4 @@ from .bench_vars import (
     OptDir,
     hash_sha1,
 )
-from .plots.plot_library import PlotLibrary, AllPlotTypes
+from .plotting.plot_library import PlotLibrary, AllPlotTypes
