+++ bencher/plotting/plot_library.py
@@ -1,6 +1,6 @@
-from bencher.plots.plot_collection import PlotCollection
-from bencher.plots.catplot import Catplot
-from bencher.plots.tables import Tables
+from bencher.plotting.plot_collection import PlotCollection
+from bencher.plotting.plots.catplot import Catplot
+from bencher.plotting.plots.tables import Tables
 
 from strenum import StrEnum
 from enum import auto
